+++ graphistry/compute/ast.py
@@ -3,7 +3,48 @@ import pandas as pd
 
 from graphistry.Plottable import Plottable
 from .predicates.ASTPredicate import ASTPredicate
-from .predicates.is_in import is_in, IsIn
+from .predicates.is_in import (
+    is_in, IsIn
+)
+from .predicates.categorical import (
+    duplicated, Duplicated,
+)
+from .predicates.temporal import (
+    is_month_start, IsMonthStart,
+    is_month_end, IsMonthEnd,
+    is_quarter_start, IsQuarterStart,
+    is_quarter_end, IsQuarterEnd,
+    is_year_start, IsYearStart,
+    is_leap_year, IsLeapYear
+)
+from .predicates.numeric import (
+    gt, GT,
+    lt, LT,
+    ge, GE,
+    le, LE,
+    eq, EQ,
+    ne, NE,
+    between, Between,
+    isna, IsNA,
+    notna, NotNA
+)
+from .predicates.str import (
+    contains, Contains,
+    startswith, Startswith,
+    endswith, Endswith,
+    match, Match,
+    isnumeric, IsNumeric,
+    isalpha, IsAlpha,
+    isdigit, IsDigit,
+    islower, IsLower,
+    isupper, IsUpper,
+    isspace, IsSpace,
+    isalnum, IsAlnum,
+    isdecimal, IsDecimal,
+    istitle, IsTitle,
+    isnull, IsNull,
+    notnull, NotNull
+)
 from .filter_by_dict import filter_by_dict
 
 import logging